requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
Werkzeug==3.0.1
python-docx==1.0.1
Pillow==10.1.0
gunicorn==21.2.0
